+Link Info Diag

+++ Drivers/sx1262/sx1262_drv.c
@@ -2101,10 +2101,10 @@ bool sx1262_init(void) {
     return res;
 }
 
-float lora_calc_data_rate(uint8_t sf_code, uint8_t bw_code, uint8_t cr_code) {
-    float data_rate = 0.0;
+double lora_calc_data_rate(uint8_t sf_code, uint8_t bw_code, uint8_t cr_code) {
+    double data_rate = 0.0;
     uint32_t bandwidth_hz = bandwidth2num((BandWidth_t)bw_code);
-    data_rate = ((float)(bandwidth_hz * sf_code * 4)) / ((powf(2.0f, (float)sf_code)) * ((float)(4 + cr_code)));
+    data_rate = ((double)(bandwidth_hz * sf_code * 4)) / ((powf(2.0f, (double)sf_code)) * ((double)(4 + cr_code)));
     return data_rate;
 }
 #ifdef HAS_DEBUG
